self code review; remove debug/pcc logs from runner
